Applied Eclipse's cleanup function on all projects with very conservative settings.
git-svn-id: https://svn.jfire.org/svn/main/trunk/org.nightlabs.editor2d.viewer.ui@9055 18c61ebf-f106-0410-9e1e-81fb82415b8f

+++ src/org/nightlabs/editor2d/viewer/ui/ZoomSupport.java
@@ -253,8 +253,8 @@ implements IZoomSupport
 	  // maybe realZoom is beyond minZoom or maxZoom
 	  double zoom = getZoom();
 
-	  double newX = (double) ((absoluteRect.x) * zoom);
-	  double newY = (double) ((absoluteRect.y) * zoom);	  	
+	  double newX = ((absoluteRect.x) * zoom);
+	  double newY = ((absoluteRect.y) * zoom);	  	
 	  	  	  	  	  
   	getViewport().setViewLocation((int)newX, (int)newY);
